+++ samples/angular-todo/src/app/app.module.ts
@@ -49,6 +49,13 @@ import { AccountIconComponent } from './icons/account-icon/account-icon.componen
 import { TodoIconComponent } from './icons/todo-icon/todo-icon.component';
 import { ActionIconComponent } from './icons/action-icon/action-icon.component';
 import { ChoiceComponent } from './features/journey/choice/choice.component';
+import { IdentityProviderComponent } from './features/journey/identityProvider/identity-provider.component';
+import { GoogleIconComponent } from './icons/google-icon/google-icon.component';
+import { AppleIconComponent } from './icons/apple-icon/apple-icon.component';
+import { WebAuthnComponent } from './features/journey/webAuthn/webAuthn.component';
+import { ConfirmationComponent } from './features/journey/confirmation/confirmation.component';
+import { TextOutputComponent } from './features/journey/textOutput/textOutput.component';
+import { FingerPrintIconComponent } from './icons/finger-print-icon/finger-print-icon.component';
 
 @NgModule({
   declarations: [
@@ -87,6 +94,13 @@ import { ChoiceComponent } from './features/journey/choice/choice.component';
     TodoIconComponent,
     ActionIconComponent,
     ChoiceComponent,
+    IdentityProviderComponent,
+    GoogleIconComponent,
+    AppleIconComponent,
+    WebAuthnComponent,
+    ConfirmationComponent,
+    TextOutputComponent,
+    FingerPrintIconComponent,
   ],
   imports: [BrowserModule, AppRoutingModule, FormsModule],
   providers: [],