+++ samples/angular-todo/src/app/features/journey/form/form.component.ts
@@ -8,12 +8,25 @@
  * of the MIT license. See the LICENSE file for details.
  */
 
-import { OnInit } from '@angular/core';
+import { EventEmitter, OnInit, Output } from '@angular/core';
 import { Component, Input } from '@angular/core';
-import { Router } from '@angular/router';
+import { ActivatedRoute, Router } from '@angular/router';
 import { environment } from '../../../../environments/environment';
-import type { FRLoginFailure, FRLoginSuccess, FRStep } from '@forgerock/javascript-sdk';
-import { FRAuth, TokenManager, UserManager } from '@forgerock/javascript-sdk';
+import type {
+  FRCallback,
+  FRLoginFailure,
+  FRLoginSuccess,
+  FRStep,
+  IdPValue,
+} from '@forgerock/javascript-sdk';
+import {
+  CallbackType,
+  FRAuth,
+  FRWebAuthn,
+  TokenManager,
+  UserManager,
+  WebAuthnStepType,
+} from '@forgerock/javascript-sdk';
 import { UserService } from '../../../services/user.service';
 
 /**
@@ -29,6 +42,10 @@ export class FormComponent implements OnInit {
    */
   @Input() action?: string;
 
+  /**
+   * the value representing whether is a webAuthn step or not
+   */
+  @Output() isWebAuthn = new EventEmitter<boolean>();
   /**
    * The current step awaiting user input and submission to AM
    */
@@ -64,9 +81,20 @@ export class FormComponent implements OnInit {
    */
   tree?: string;
 
-  constructor(private router: Router, public userService: UserService) {}
+  identityProviders: IdPValue[];
+  code: string;
+  state: string;
+  showWebAuthn = false;
+  webAuthnType: WebAuthnStepType;
 
+  constructor(
+    private router: Router,
+    public userService: UserService,
+    private route: ActivatedRoute,
+  ) {}
   ngOnInit(): void {
+    this.code = this.route.snapshot.queryParamMap.get('code');
+    this.state = this.route.snapshot.queryParamMap.get('state');
     this.setConfigForAction(this.action);
     this.nextStep();
   }
@@ -81,12 +109,39 @@ export class FormComponent implements OnInit {
     try {
       /** *********************************************************************
        * SDK INTEGRATION POINT
+       * Summary: Handle selection of IdpCallbacks
+       * ----------------------------------------------------------------------
+       * Details: If we do not initially select the IdpCallback, when pressing enter,
+       * local authentication will automatically be submitted.
+       * This allows us to submit the form with the proper selection of an
+       * identity provider rather than local authentication.
+       ********************************************************************* */
+
+      const selectIdPCallback = step?.getCallbacksOfType(CallbackType.SelectIdPCallback);
+
+      if (selectIdPCallback?.length > 0) {
+        const nameCallBacksInputs = step?.getCallbackOfType(CallbackType.NameCallback);
+        const idToken2Input = nameCallBacksInputs?.getInputValue('IDToken2');
+
+        if (this.isIdentityProviderLogin(selectIdPCallback[0]) && idToken2Input !== '') {
+          selectIdPCallback[0].setInputValue('localAuthentication', 0);
+        }
+      }
+
+      /** *********************************************************************
+       * SDK INTEGRATION POINT
        * Summary: Call the SDK's next method to submit the current step.
        * ----------------------------------------------------------------------
        * Details: This calls the next method with the previous step, expecting
        * the next step to be returned, or a success or failure.
        ********************************************************************* */
-      const nextStep = await FRAuth.next(step, { tree: this.tree });
+
+      let nextStep;
+      if (this.code && this.state) {
+        nextStep = await FRAuth.resume(window.location.href);
+      } else {
+        nextStep = await FRAuth.next(step, { tree: this.tree });
+      }
 
       /** *******************************************************************
        * SDK INTEGRATION POINT
@@ -162,6 +217,21 @@ export class FormComponent implements OnInit {
   handleStep(step?: FRStep) {
     this.step = step;
 
+    /**
+     * Since WebAuthn is being handled as its own unique component, it is necessary to check if we have WebAuthn type step.
+     * if so an output event is emitted to the log-in component to switch the login icon and render the webAuhtn component into the form component.
+     */
+    const webAuthnType = FRWebAuthn.getWebAuthnStepType(step);
+    if (webAuthnType !== WebAuthnStepType.None) {
+      this.showWebAuthn = true;
+      this.webAuthnType = webAuthnType;
+      this.isWebAuthn.emit(true);
+    }
+
+    const redirectCallback = step?.getCallbacksOfType(CallbackType.RedirectCallback);
+    if (redirectCallback?.length > 0) {
+      FRAuth.redirect(step);
+    }
     this.setConfigForAction(this.action);
 
     if (step?.getHeader()) {
@@ -177,13 +247,13 @@ export class FormComponent implements OnInit {
     switch (action) {
       case 'login': {
         this.title = 'Sign In';
-        this.buttonText = 'Sign In';
+        this.buttonText = 'Submit';
         this.tree = environment.JOURNEY_LOGIN;
         break;
       }
       case 'register': {
         this.title = 'Sign Up';
-        (this.buttonText = 'Register'), (this.tree = environment.JOURNEY_REGISTER);
+        (this.buttonText = 'Submit'), (this.tree = environment.JOURNEY_REGISTER);
         break;
       }
       default: {
@@ -193,5 +263,24 @@ export class FormComponent implements OnInit {
         break;
       }
     }
+  }
+
+  isIdentityProviderLogin(callback: FRCallback): boolean {
+    return callback
+      ?.getOutputByName('providers', [])
+      .filter((provider) => provider.provider !== 'localAuthentication').length > 0
+      ? true
+      : false;
+  }
+
+  requiresUserInput(callbacks: FRCallback[]): boolean {
+    return callbacks.filter((callback) => {
+      return (
+        callback.getType() === CallbackType.ConfirmationCallback ||
+        callback.getType() === CallbackType.SelectIdPCallback
+      );
+    }).length > 0
+      ? false
+      : true;
   }
 }