+++ samples/angular-todo/src/app/shared/idps.types.ts
@@ -1,0 +1,16 @@
+/*
+ * @forgerock/javascript-sdk
+ *
+ * idps.types.ts
+ *
+ * Copyright (c) 2020 ForgeRock. All rights reserved.
+ * This software may be modified and distributed under the terms
+ * of the MIT license. See the LICENSE file for details.
+ */
+
+/**
+ * Types of identity providers
+ */
+type IdpNames = 'Google' | 'Apple' | 'Facebook';
+
+export { IdpNames };