+++ samples/angular-todo/src/app/views/login/login.component.ts
@@ -17,4 +17,10 @@ import { Component } from '@angular/core';
   selector: 'app-login',
   templateUrl: './login.component.html',
 })
-export class LoginComponent {}
+export class LoginComponent {
+  isWebAuthn = false;
+
+  onSetIsWebAuthn(isWebAuthn: boolean): void {
+    this.isWebAuthn = isWebAuthn;
+  }
+}